version 2.3 Images added for 2048 and memory game and organized files
