+++ .idea/misc.xml
@@ -3,20 +3,16 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/GitHub/K-Dev/app/src/main/res/layout/activity_main_2048game.xml" value="0.17806484295845998" />
-        <entry key="..\:/GitHub/K-Dev/app/src/main/res/layout/activity_show_team_info.xml" value="0.359375" />
-        <entry key="..\:/GitHub/K-Dev/app/src/main/res/layout/game_2048.xml" value="0.17477203647416414" />
-        <entry key="..\:/GitHub/K-Dev/app/src/main/res/layout/instruction_2048.xml" value="0.16666666666666666" />
-        <entry key="..\:/GitHub/K-Dev/app/src/main/res/layout/popup_instruction_2048.xml" value="0.14666666666666667" />
-        <entry key="..\:/GitHub/K-Dev/app/src/main/res/menu/popup_menu.xml" value="0.36614583333333334" />
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.1651469098277609" />
-        <entry key="app/src/main/res/layout/activity_main_2048game.xml" value="0.4332955819589121" />
-        <entry key="app/src/main/res/layout/activity_memorygame.xml" value="0.335" />
-        <entry key="app/src/main/res/layout/game_2048.xml" value="0.1" />
-        <entry key="app/src/main/res/layout/item_card.xml" value="0.19522776572668113" />
-        <entry key="app/src/main/res/menu/popup_2048.xml" value="0.2916666666666667" />
+        <entry key="app/src/main/res/layout/frame_1.xml" value="0.16394927536231885" />
+        <entry key="app/src/main/res/layout/frame_2.xml" value="0.16394927536231885" />
+        <entry key="app/src/main/res/layout/frame_3.xml" value="0.16394927536231885" />
+        <entry key="app/src/main/res/layout/game_2048.xml" value="0.1703125" />
+        <entry key="app/src/main/res/layout/item_card.xml" value="0.16394927536231885" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK" />
+  <component name="ExternalStorageConfigurationManager" enabled="true" />
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android API 32 Platform" project-jdk-type="Android SDK">
+    <output url="file://$PROJECT_DIR$/out" />
+  </component>
 </project>
