+++ app/src/main/AndroidManifest.xml
@@ -33,12 +33,12 @@
         </activity>
 
         <activity
-            android:name=".LogicGame2048" android:screenOrientation="portrait"
+            android:name=".game2048.LogicGame2048" android:screenOrientation="portrait"
             android:exported="true">
 
         </activity>
         <activity
-            android:name=".MemoryGame" android:screenOrientation="portrait"
+            android:name=".memorygame.MemoryGame" android:screenOrientation="portrait"
             android:exported="true">
 
         </activity>
