+++ app/src/main/java/com/example/k_dev_master/Fragment_1.java
@@ -4,11 +4,9 @@ import static com.example.k_dev_master.MainActivity.gameSelect;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.PopupMenu;
-import android.widget.Toast;
+
 
 import androidx.fragment.app.Fragment;
 
@@ -17,7 +15,7 @@ public class Fragment_1 extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.game_2048, container, false);
+                R.layout.frame_1, container, false);
         gameSelect = 1;
         return rootView;
     }
